move BearerTokenCredentialsPolicy to _internal
